refactor(taint): split 1526-line TaintAnalyzer into 4 language strategies
Extract language-specific taint analysis into strategy classes:
- PythonTaintStrategy (547 lines) — AST-based interprocedural analysis
- JsTaintStrategy (163 lines) — regex-based JS/TS analysis
- GoTaintStrategy (108 lines) — regex-based Go analysis
- JavaTaintStrategy (107 lines) — regex-based Java analysis

TaintAnalyzer reduced from 1526 to 554 lines (shared types + dispatch).
475 taint tests pass unchanged.
